+++ app/components/profile/terms-of-use/TermsOfUseForm.js
@@ -2,6 +2,7 @@
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
 import Button from 'react-toolbox/lib/button/Button';
+import ReactMarkdown from 'react-markdown';
 import { defineMessages, intlShape } from 'react-intl';
 import CheckboxWithLongLabel from '../../widgets/forms/CheckboxWithLongLabel';
 import LocalizableError from '../../../i18n/LocalizableError';
@@ -24,6 +25,7 @@ const messages = defineMessages({
 export default class TermsOfUseForm extends Component {
 
   props: {
+    localizedTermsOfUse: string,
     onSubmit: Function,
     isSubmitting: boolean,
     error?: ?LocalizableError,
@@ -47,7 +49,7 @@ export default class TermsOfUseForm extends Component {
 
   render() {
     const { intl } = this.context;
-    const { isSubmitting, error } = this.props;
+    const { isSubmitting, error, localizedTermsOfUse } = this.props;
     const { areTermsOfUseAccepted } = this.state;
 
     return (
@@ -55,61 +57,8 @@ export default class TermsOfUseForm extends Component {
         <div className={styles.centeredBox}>
 
           <div className={styles.terms}>
-            <h1>Terms of use</h1>
 
-            <p>
-              First paragraph quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur,
-              adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore
-              some link aliquam quaerat voluptatem.
-            </p>
-
-            <p>
-              Second paragraph ad minima veniam, quis nostrum exercitationem ullam corporis
-              suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum
-              iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur,
-              vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?
-            </p>
-
-            <ol>
-              <li>
-                <h2>1. Numbered title</h2>
-                <ul>
-                  <li>
-                    <strong>1.1 Sub-title name.</strong>
-                    At vero eos et accusamus et iusto odio dignissimos ducimus qui blanditiis
-                    praesentium voluptatum deleniti atque corrupti quos dolores et quas molestias
-                    excepturi sint occaecati cupiditate non provident, similique sunt in culpa qui
-                    officia deserunt mollitia animi, id est laborum et dolorum fuga.
-                  </li>
-                  <li>
-                    <strong>1.2 Sub-title name.</strong>
-                    Et harum quidem rerum facilis est et expedita distinctio. Nam libero tempore,
-                    cum soluta nobis est eligendi optio cumque nihil impedit quo minus id quod
-                    maxime placeat facere possimus, omnis voluptas assumenda est,
-                    omnis dolor repellendus.
-                  </li>
-                </ul>
-              </li>
-
-              <li>
-                <h2>2. Numbered title</h2>
-                <ul>
-                  <li>
-                    <strong>2.1 Sub-title name.</strong>
-                    At vero eos et accusamus et iusto odio dignissimos ducimus qui blanditiis
-                    praesentium voluptatum deleniti atque corrupti quos dolores et quas molestias
-                    excepturi sint occaecati cupiditate non provident, similique sunt in culpa...
-                    and so on...
-                  </li>
-                </ul>
-              </li>
-            </ol>
-
-            <p>
-              Last paragraph quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur,
-              adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore
-              some link aliquam quaerat voluptatem.
-            </p>
+            <ReactMarkdown source={localizedTermsOfUse} />
 
             <CheckboxWithLongLabel
               className={styles.checkbox}